requirements.txt
backcall==0.2.0
bokeh==2.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
colorcet==2.0.6
cycler==0.10.0
dask==2021.3.1
datashader==0.12.1
datashape==0.5.2
decorator==4.4.2
distributed==2021.3.1
fsspec==0.8.7
HeapDict==1.0.1
holoviews==1.14.2
idna==2.10
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
llvmlite==0.36.0
locket==0.2.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.0
msgpack==1.0.2
multipledispatch==0.6.0
numba==0.53.1
numpy==1.20.2
packaging==20.9
pandas==1.2.3
panel==0.11.1
param==1.10.1
parso==0.8.1
partd==1.1.0
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.18
psutil==5.8.0
pyct==0.4.8
Pygments==2.8.1
pynndescent==0.5.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pyviz-comms==2.0.1
PyYAML==5.4.1
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
six==1.15.0
sklearn==0.0
sortedcontainers==2.3.0
tblib==1.7.0
threadpoolctl==2.1.0
toolz==0.11.1
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typing-extensions==3.7.4.3
umap-learn==0.5.1
urllib3==1.26.4
wcwidth==0.2.5
xarray==0.17.0
zict==2.0.0
